dominio movimiento hecho. Listo para probar.

+++ src/main/java/com/devsu/challenge/backend/apirest/accountapp/controller/MovimientoController.java
@@ -1,9 +1,12 @@
 package com.devsu.challenge.backend.apirest.accountapp.controller;
 
 import com.devsu.challenge.backend.apirest.accountapp.entity.Movimiento;
+import com.devsu.challenge.backend.apirest.accountapp.exceptions.CuentaNoExisteException;
 import com.devsu.challenge.backend.apirest.accountapp.service.IMovimientoService;
 import java.util.List;
 import java.util.Optional;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -24,28 +27,55 @@ public class MovimientoController {
     }
 
     @GetMapping
-    public List<Movimiento> getAllMovimientos() {
-        return movimientoService.getAllMovimientos();
+    public ResponseEntity<List<Movimiento>> getAllMovimientos() {
+        List<Movimiento> movimientos = movimientoService.getAllMovimientos();
+        if (movimientos.isEmpty()) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<>(movimientos, HttpStatus.OK);
     }
 
     @PostMapping
-    public Movimiento createMovimiento(@RequestBody Movimiento movimiento) {
-        return movimientoService.createMovimiento(movimiento);
+    public ResponseEntity<Movimiento> crearMovimiento(@RequestBody Movimiento movimiento) {
+        try {
+            Movimiento nuevoMov = movimientoService.createMovimiento(movimiento);
+            return new ResponseEntity<>(nuevoMov, HttpStatus.CREATED);
+        } catch (CuentaNoExisteException ex) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        } catch (Exception e) {
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+
     }
 
     @GetMapping("/{id}")
-    public Optional<Movimiento> getMovimientoById(@PathVariable Long id) {
-        return movimientoService.getMovimientoById(id);
+    public ResponseEntity<Movimiento> getMovimientoById(@PathVariable Long id) {
+        Optional<Movimiento> movimiento = movimientoService.getMovimientoById(id);
+        return movimiento
+            .map(mov -> new ResponseEntity<>(mov, HttpStatus.OK))
+            .orElseGet(() -> new ResponseEntity<>(HttpStatus.NOT_FOUND));
     }
 
     @PutMapping("/{id}")
-    public Movimiento updateMovimiento(@PathVariable Long id, @RequestBody Movimiento movimientoDetails) {
-        return movimientoService.updateMovimiento(id, movimientoDetails);
+    public ResponseEntity<Movimiento> updateMovimiento(@PathVariable Long id, @RequestBody Movimiento movimientoDetalle) {
+        Movimiento updatedMov = movimientoService.updateMovimiento(id, movimientoDetalle);
+        if (updatedMov == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<>(updatedMov, HttpStatus.OK);
     }
 
     @DeleteMapping("/{id}")
-    public void deleteMovimiento(@PathVariable Long id) {
-        movimientoService.deleteMovimiento(id);
+    public ResponseEntity<Void> deleteMovimiento(@PathVariable Long id) {
+        Optional<Movimiento> movimiento = movimientoService.getMovimientoById(id);
+
+        if (movimiento.isPresent()) {
+            movimientoService.deleteMovimiento(id);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
     }
 
 }
